fix(project): harden KAM check and use safe Content-Disposition builder
- Cast both sides of KAM comparison to int to avoid silent denial if PDO
  config ever returns IDs as strings.
- Drop hand-built Content-Disposition header; rely on FilesystemAdapter
  default which routes filename through HeaderUtils::makeDisposition()
  (RFC 5987 safe encoding for non-ASCII/quotes/CRLF in filenames).

+++ Modules/Project/Policies/ProjectContractPolicy.php
@@ -21,7 +21,7 @@ class ProjectContractPolicy
             return false;
         }
 
-        $isKAM = optional($project->client)->key_account_manager_id === $user->id;
+        $isKAM = (int) optional($project->client)->key_account_manager_id === (int) $user->id;
         if ($isKAM) {
             return true;
         }
